Make ElectrodeNames private
