+++ pulse2percept/implants/electrode_arrays.py
@@ -10,7 +10,7 @@ from skimage.transform import SimilarityTransform
 from copy import deepcopy
 
 from .electrodes import Electrode, PointSource, HexElectrode
-from ..stimuli.base import ElectrodeNames
+from ..stimuli._grid_names import _GridNames
 from ..units import Quantity, as_value, deg, um
 from ..utils import PrettyPrint, bijective26_name
 from ..utils.constants import ZORDER
@@ -23,7 +23,7 @@ def _is_electrode_collection(selector):
     ``(row, col)`` pair on an
     :py:class:`~pulse2percept.implants.ElectrodeGrid`, and an index. Anything
     else that can be iterated is a collection -- a list, an array, or the
-    :py:class:`~pulse2percept.stimuli.ElectrodeNames` a stimulus reports.
+    name container a stimulus reports as ``electrodes``.
 
     The tuple carve-out is what makes ``grid[0, 0]`` and
     ``grid.coordinates(electrodes=(0, 0))`` mean the same thing, and the
@@ -441,10 +441,9 @@ class ElectrodeGrid(ElectrodeArray):
         For example ('1', 'A') will number rows numerically and columns
         alphabetically; first row: 'A1', 'B1', 'C1', NOT '1A', '1B', '1C'.
 
-        The default, ``('A', '1')``, is the same convention that
-        :py:class:`~pulse2percept.stimuli.ElectrodeNames` uses to name the
+        The default, ``('A', '1')``, is the same convention that names the
         pixels of an :py:class:`~pulse2percept.stimuli.ImageStimulus`, and is
-        generated by it. The other combinations exist to reproduce the naming
+        generated by the same implementation. The other combinations exist to reproduce the naming
         of specific published implants and are not otherwise recommended.
 
         Alternatively, pass a list or NumPy array with one name per electrode
@@ -725,7 +724,7 @@ class ElectrodeGrid(ElectrodeArray):
                 # The remaining schemes above and below exist to reproduce the
                 # naming of specific published implants (ArgusI, Orion), and
                 # are deliberately left as overrides:
-                names = np.asarray(ElectrodeNames((rows, cols))).tolist()
+                names = np.asarray(_GridNames((rows, cols))).tolist()
             else:
                 names = [rws[i] + clms[j] for i in range(len(rws))
                          for j in range(len(clms))]
